+++ src/provider/semanticTokensRangeManager.ts
@@ -31,7 +31,6 @@ export default class SemanticTokensRangeManager extends Manager<DocumentRangeSem
     if (!item) return null
     let { provider } = item
     if (provider.provideDocumentRangeSemanticTokens === null) return null
-
     return await Promise.resolve(provider.provideDocumentRangeSemanticTokens(document, range, token))
   }
 }